Reorganize blob interface
